move InternalLogger duplicates to .tt
