+++ src/NLog/Common/InternalLogger.cs
@@ -51,7 +51,7 @@ namespace NLog.Common
     /// <summary>
     /// NLog internal logger.
     /// </summary>
-    public static class InternalLogger
+    public static partial class InternalLogger
     {
         private static object lockObject = new object();
         private static string _logFile;
@@ -125,54 +125,6 @@ namespace NLog.Common
         public static bool IncludeTimestamp { get; set; }
 
         /// <summary>
-        /// Gets a value indicating whether internal log includes Trace messages.
-        /// </summary>
-        public static bool IsTraceEnabled
-        {
-            get { return LogLevel.Trace >= LogLevel; }
-        }
-
-        /// <summary>
-        /// Gets a value indicating whether internal log includes Debug messages.
-        /// </summary>
-        public static bool IsDebugEnabled
-        {
-            get { return LogLevel.Debug >= LogLevel; }
-        }
-
-        /// <summary>
-        /// Gets a value indicating whether internal log includes Info messages.
-        /// </summary>
-        public static bool IsInfoEnabled
-        {
-            get { return LogLevel.Info >= LogLevel; }
-        }
-
-        /// <summary>
-        /// Gets a value indicating whether internal log includes Warn messages.
-        /// </summary>
-        public static bool IsWarnEnabled
-        {
-            get { return LogLevel.Warn >= LogLevel; }
-        }
-
-        /// <summary>
-        /// Gets a value indicating whether internal log includes Error messages.
-        /// </summary>
-        public static bool IsErrorEnabled
-        {
-            get { return LogLevel.Error >= LogLevel; }
-        }
-
-        /// <summary>
-        /// Gets a value indicating whether internal log includes Fatal messages.
-        /// </summary>
-        public static bool IsFatalEnabled
-        {
-            get { return LogLevel.Fatal >= LogLevel; }
-        }
-
-        /// <summary>
         /// Logs the specified message at the specified level.
         /// </summary>
         /// <param name="level">Log level.</param>
@@ -192,126 +144,6 @@ namespace NLog.Common
         public static void Log(LogLevel level, [Localizable(false)] string message)
         {
             Write(level, message, null);
-        }
-
-        /// <summary>
-        /// Logs the specified message at the Trace level.
-        /// </summary>
-        /// <param name="message">Message which may include positional parameters.</param>
-        /// <param name="args">Arguments to the message.</param>
-        [StringFormatMethod("message")]
-        public static void Trace([Localizable(false)] string message, params object[] args)
-        {
-            Write(LogLevel.Trace, message, args);
-        }
-
-        /// <summary>
-        /// Logs the specified message at the Trace level.
-        /// </summary>
-        /// <param name="message">Log message.</param>
-        public static void Trace([Localizable(false)] string message)
-        {
-            Write(LogLevel.Trace, message, null);
-        }
-
-        /// <summary>
-        /// Logs the specified message at the Debug level.
-        /// </summary>
-        /// <param name="message">Message which may include positional parameters.</param>
-        /// <param name="args">Arguments to the message.</param>
-        [StringFormatMethod("message")]
-        public static void Debug([Localizable(false)] string message, params object[] args)
-        {
-            Write(LogLevel.Debug, message, args);
-        }
-
-        /// <summary>
-        /// Logs the specified message at the Debug level.
-        /// </summary>
-        /// <param name="message">Log message.</param>
-        public static void Debug([Localizable(false)] string message)
-        {
-            Write(LogLevel.Debug, message, null);
-        }
-
-        /// <summary>
-        /// Logs the specified message at the Info level.
-        /// </summary>
-        /// <param name="message">Message which may include positional parameters.</param>
-        /// <param name="args">Arguments to the message.</param>
-        [StringFormatMethod("message")]
-        public static void Info([Localizable(false)] string message, params object[] args)
-        {
-            Write(LogLevel.Info, message, args);
-        }
-
-        /// <summary>
-        /// Logs the specified message at the Info level.
-        /// </summary>
-        /// <param name="message">Log message.</param>
-        public static void Info([Localizable(false)] string message)
-        {
-            Write(LogLevel.Info, message, null);
-        }
-
-        /// <summary>
-        /// Logs the specified message at the Warn level.
-        /// </summary>
-        /// <param name="message">Message which may include positional parameters.</param>
-        /// <param name="args">Arguments to the message.</param>
-        [StringFormatMethod("message")]
-        public static void Warn([Localizable(false)] string message, params object[] args)
-        {
-            Write(LogLevel.Warn, message, args);
-        }
-
-        /// <summary>
-        /// Logs the specified message at the Warn level.
-        /// </summary>
-        /// <param name="message">Log message.</param>
-        public static void Warn([Localizable(false)] string message)
-        {
-            Write(LogLevel.Warn, message, null);
-        }
-
-        /// <summary>
-        /// Logs the specified message at the Error level.
-        /// </summary>
-        /// <param name="message">Message which may include positional parameters.</param>
-        /// <param name="args">Arguments to the message.</param>
-        [StringFormatMethod("message")]
-        public static void Error([Localizable(false)] string message, params object[] args)
-        {
-            Write(LogLevel.Error, message, args);
-        }
-
-        /// <summary>
-        /// Logs the specified message at the Error level.
-        /// </summary>
-        /// <param name="message">Log message.</param>
-        public static void Error([Localizable(false)] string message)
-        {
-            Write(LogLevel.Error, message, null);
-        }
-
-        /// <summary>
-        /// Logs the specified message at the Fatal level.
-        /// </summary>
-        /// <param name="message">Message which may include positional parameters.</param>
-        /// <param name="args">Arguments to the message.</param>
-        [StringFormatMethod("message")]
-        public static void Fatal([Localizable(false)] string message, params object[] args)
-        {
-            Write(LogLevel.Fatal, message, args);
-        }
-
-        /// <summary>
-        /// Logs the specified message at the Fatal level.
-        /// </summary>
-        /// <param name="message">Log message.</param>
-        public static void Fatal([Localizable(false)] string message)
-        {
-            Write(LogLevel.Fatal, message, null);
         }
 
         private static void Write(LogLevel level, string message, object[] args)
