+++ src/NLog/Common/InternalLogger.tt
@@ -1,0 +1,86 @@
+// 
+// Copyright (c) 2004-2011 Jaroslaw Kowalski <jaak@jkowalski.net>
+// 
+// All rights reserved.
+// 
+// Redistribution and use in source and binary forms, with or without 
+// modification, are permitted provided that the following conditions 
+// are met:
+// 
+// * Redistributions of source code must retain the above copyright notice, 
+//   this list of conditions and the following disclaimer. 
+// 
+// * Redistributions in binary form must reproduce the above copyright notice,
+//   this list of conditions and the following disclaimer in the documentation
+//   and/or other materials provided with the distribution. 
+// 
+// * Neither the name of Jaroslaw Kowalski nor the names of its 
+//   contributors may be used to endorse or promote products derived from this
+//   software without specific prior written permission. 
+// 
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE 
+// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE 
+// ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE 
+// LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR 
+// CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+// SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS 
+// INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN 
+// CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) 
+// ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF 
+// THE POSSIBILITY OF SUCH DAMAGE.
+// 
+
+namespace NLog.Common
+{
+    using JetBrains.Annotations;
+    using System;
+    using System.ComponentModel;
+	
+    public static partial class InternalLogger
+    {
+        
+ <#
+    var levels = new string[]{"Trace", "Debug", "Info", "Warn", "Error", "Fatal"};
+
+    foreach(var level in levels)
+    { 
+#>
+        /// <summary>
+        /// Gets a value indicating whether internal log includes <#=level#> messages.
+        /// </summary>
+        public static bool Is<#=level#>Enabled
+        {
+            get { return LogLevel.<#=level#> >= LogLevel; }
+        }
+
+<#
+    }
+	   foreach(var level in levels)
+    { 
+#>
+
+        /// <summary>
+        /// Logs the specified message at the <#=level#> level.
+        /// </summary>
+        /// <param name="message">Message which may include positional parameters.</param>
+        /// <param name="args">Arguments to the message.</param>
+        [StringFormatMethod("message")]
+        public static void <#=level#>([Localizable(false)] string message, params object[] args)
+        {
+            Write(LogLevel.<#=level#>, message, args);
+        }
+
+        /// <summary>
+        /// Logs the specified message at the <#=level#> level.
+        /// </summary>
+        /// <param name="message">Log message.</param>
+        public static void <#=level#>([Localizable(false)] string message)
+        {
+            Write(LogLevel.<#=level#>, message, null);
+        }
+<#
+	}
+ #>     
+    }
+}
